remove the circular buffer, add manually implemented bulk IN transfers
All the circular buffer packetization code occured to be too slow on
this microcontroller. CDC code is not used anymore. Raw bulk transfers
are used instead. All the ADC and encoders (only lower 16 bit!) values
will be written directly to the buffer of the IN endpoint 1.
Each time the host tries to read it the whole buffer will be returned.
Thus each value in the buffer will be either the most recent one or
a little bit outdated. We rely on atomicity of 16-bit RAM write operations
here, so the full 32-bit encoder values will be calculated by the host.

The ADC values are supposed to be copied by the DMA from ADCMEM0 to the
endpoint buffer.

The encoder values are supposed to be incremented by the CPU (mod 2^16)
in the interrupt handler.

Commands from the host are supposed to be sent one command per transfer,
so that we could parse them fast and easy here.