a lot of changes (see below)
It seems that the USB API does not initialize data endpoints anymore, because the number of interfaces is set to 0.
So stUsbHandle array is empty and useless now, we must initialize the endpoints manually instead (see main.c).

In descriptors:
* Device class is changed to "vendor specific" so that host OS wouldn't try to load any drivers.
  So no CDC requests are expected	from the host. Only the	basic USB control requests (we hope they will be handled by the	USB API).
* Device strings are changed.
* VID still remains to be "Texas instruments"
* Only one vendor-specific interface remains with bulk endpoints 1
* the abromConfigurationDescriptorGroup is radically simplified Cypress-style (a char array)
* USB_SUPPORT_SELF_POWERED changed, because our device is self-powered

INT_EP macroses were changed to show that we only use endpoints 0 and 1.